From long to long long all the way

+++ dlib/dnn/tensor_abstract.h
@@ -46,7 +46,7 @@ namespace dlib
 
         virtual ~tensor();
 
-        long num_samples(
+        long long num_samples(
         ) const; 
         /*!
             ensures
@@ -54,7 +54,7 @@ namespace dlib
                   are in this object.  
         !*/
 
-        long k(
+        long long k(
         ) const; 
         /*!
             ensures
@@ -63,14 +63,14 @@ namespace dlib
                   with k() channels.
         !*/
 
-        long nr(
+        long long nr(
         ) const; 
         /*!
             ensures
                 - returns the number of rows in this tensor.
         !*/
 
-        long nc(
+        long long nc(
         ) const; 
         /*!
             ensures
@@ -288,7 +288,7 @@ namespace dlib
 
         template <typename EXP>
         void set_sample (
-            unsigned long idx,
+            unsigned long long idx,
             const matrix_exp<EXP>& item
         );
         /*!
@@ -304,7 +304,7 @@ namespace dlib
 
         template <typename EXP>
         void add_to_sample (
-            unsigned long idx,
+            unsigned long long idx,
             const matrix_exp<EXP>& item
         );
         /*!
@@ -363,8 +363,8 @@ namespace dlib
 
     const matrix_exp mat (
         const tensor& t,
-        long nr,
-        long nc
+        long long nr,
+        long long nc
     );
     /*!
         requires
@@ -394,8 +394,8 @@ namespace dlib
 
     const matrix_exp image_plane (
         const tensor& t,
-        long sample = 0,
-        long k = 0
+        long long sample = 0,
+        long long k = 0
     );
     /*!
         requires
@@ -467,7 +467,7 @@ namespace dlib
         !*/
 
         explicit resizable_tensor(
-            long n_, long k_ = 1, long nr_ = 1, long nc_ = 1
+            long long n_, long long k_ = 1, long long nr_ = 1, long long nc_ = 1
         );
         /*!
             requires
@@ -524,7 +524,7 @@ namespace dlib
         !*/
 
         void set_size(
-            long n_, long k_ = 1, long nr_ = 1, long nc_ = 1
+            long long n_, long long k_ = 1, long long nr_ = 1, long long nc_ = 1
         );
         /*!
             requires
@@ -656,7 +656,7 @@ namespace dlib
         !*/
 
         alias_tensor (
-            long n_, long k_ = 1, long nr_ = 1, long nc_ = 1
+            long long n_, long long k_ = 1, long long nr_ = 1, long long nc_ = 1
         );
         /*!
             requires
@@ -672,10 +672,10 @@ namespace dlib
                 - #nc() == nc_
         !*/
 
-        long num_samples() const; 
-        long k() const; 
-        long nr() const; 
-        long nc() const; 
+        long long num_samples() const;
+        long long k() const;
+        long long nr() const;
+        long long nc() const;
         size_t size() const;
 
         alias_tensor_instance operator() (
